ref: move components to components folder
