+++ app/components/FloatingBubbles.tsx
